[Refactor][Joshua] Restructured screen folders
